dom/datastore/tests/test_app_install.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for DataStore - install/uninstall apps</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript;version=1.7">

  SimpleTest.waitForExplicitFinish();

  var gBaseURL = 'http://test/tests/dom/datastore/tests/';
  var gHostedManifestURL = gBaseURL + 'file_app.sjs';
  var gGenerator = runTest();

  function go() {
    SpecialPowers.pushPermissions(
      [{ "type": "browser", "allow": 1, "context": document },
       { "type": "embed-apps", "allow": 1, "context": document },
       { "type": "webapps-manage", "allow": 1, "context": document }],
      function() { gGenerator.next() });
  }

  function continueTest() {
    gGenerator.next();
  }

  function cbError() {
    ok(false, "Error callback invoked");
    finish();
  }

  function runTest() {
    ok("getDataStores" in navigator, "getDataStores exists");
    is(typeof navigator.getDataStores, "function", "getDataStores exists and it's a function");
    navigator.getDataStores('foo').then(function(stores) {
      is(stores.length, 0, "getDataStores('foo') returns 0 elements");
      continueTest();
    }, cbError);
    yield undefined;

    SpecialPowers.autoConfirmAppInstall(continueTest);
    yield undefined;

    var request = navigator.mozApps.install(gHostedManifestURL);
    request.onerror = cbError;
    request.onsuccess = continueTest;
    yield undefined;

    var app = request.result;
    isnot(app, null, "App is non-null");
    is(app.manifest.description, "Updated even faster than Firefox, just to annoy slashdotters.",
       "Manifest is HTML-sanitized");

    navigator.getDataStores('foo').then(function(stores) {
      is(stores.length, 1, "getDataStores('foo') returns 1 element");
      is(stores[0].name, 'foo', 'The dataStore.name is foo');
      is(stores[0].owner, 'http://test/tests/dom/datastore/tests/file_app.sjs', 'The dataStore.owner exists');
      is(stores[0].readOnly, false, 'The dataStore foo is not in readonly');
      continueTest();
    }, cbError);
    yield undefined;

    navigator.getDataStores('bar').then(function(stores) {
      is(stores.length, 1, "getDataStores('bar') returns 1 element");
      is(stores[0].name, 'bar', 'The dataStore.name is bar');
      is(stores[0].owner, 'http://test/tests/dom/datastore/tests/file_app.sjs', 'The dataStore.owner exists');
      is(stores[0].readOnly, true, 'The dataStore bar is in readonly');
      continueTest();
    }, cbError);
    yield undefined;

    // Uninstall the app.
    request = navigator.mozApps.mgmt.uninstall(app);
    request.onerror = cbError;
    request.onsuccess = continueTest;
    yield undefined;

    navigator.getDataStores('foo').then(function(stores) {
      is(stores.length, 0, "getDataStores('foo') returns 0 elements");
      continueTest();
    }, cbError);
    yield undefined;

    // All done.
    info("All done");
    finish();
  }

  function finish() {
    SimpleTest.finish();
  }

  </script>
</head>
<body onload="go()">
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
<div id="container"></div>
</body>
</html>
